add verification of books.json to CI (#59)
This PR adds a verification of the file `books.json` (including the SRI checksums) to the CI.

The workflow takes about 15 minutes, because it verifies _all_ the statistics stored in `books.json` as well.

If we just want to check the SRI's, then this can be done in under a minute. But I think it is best to check the whole file for correctness.

While at it, we add a missing newline character to the JSON file. Gemini says `This adheres to POSIX standards where text files should end with a newline.`